services/news-signal/pyproject.toml
[project]
name = "news-signal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index-core>=0.12.8",
    "llama-index-llms-openai>=0.3.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.6.1",
    "quixstreams>=3.5.0",
]
